nds/sound: Add soundSetMixerSleep, see motivation:
When the NDS mixer is disabled (REG_SOUNDCNT.bit15=0), playback state
for all active channels is reset. When the mixer is turned back on,
all active channels are restarted from the beginning. This causes
problems when implementing timing critical playback (such as ring
buffer sound streaming).

This commit adds a function which allows apps to control whether
the NDS mixer is actually disabled (or left enabled) when sleep mode
is triggered by e.g. closing the lid.